+++ src/core/server/api/consts/constSystem.ts
@@ -9,12 +9,14 @@ import { PluginSystem } from '@AthenaServer/systems/plugins';
 import { SoundSystem } from '@AthenaServer/systems/sound';
 import { defaultSystemsConst } from './constDefaultSystems';
 import { ItemManager } from '@AthenaServer/systems/itemManager';
+import { ItemCrafting } from '@AthenaServer/systems/itemCrafting';
 
 export const systemConst = {
     agenda: AgendaSystem,
     default: defaultSystemsConst,
     itemFactory: ItemFactory,
     itemManager: ItemManager,
+    itemCrafting: ItemCrafting,
     effects: ItemEffects,
     identifier: Identifier,
     job: {